Standardize package.json scripts (#15)
* Standardize package.json scripts

This brings the package.json scripts inline with the standard names we
use for all other Dexa repos. It also updates the GitHub workflow for
the changes and to use explicit jobs for nicer reporting.

* Fix duplicate GitHub workflow run

* Stop running test on multiple Node versions

It's noisy, unlikely to catch anything with unit tests, and we
exclusively use Node v18 so others can open PRs for issues with other
versions.