Add OpenTelemetry tracing via Sentry
This adds OpenTelemetry tracing via Sentry to the base model class,
which applies it to all models.

It currently makes Sentry a peer dependency, which means we either need
to do a major version bump and force all users into Sentry, or find a
way to conditionally import Sentry and handle the case where it's not
installed.